+++ test/groovy/SetupCommonPipelineEnvironmentTest.groovy
@@ -15,6 +15,7 @@ import util.Rules
 
 import static org.junit.Assert.assertEquals
 import static org.junit.Assert.assertNotNull
+import static org.junit.Assert.assertNull
 
 class SetupCommonPipelineEnvironmentTest extends BasePiperTest {
 
@@ -206,5 +207,43 @@ class SetupCommonPipelineEnvironmentTest extends BasePiperTest {
         assertEquals("custom: 'myRemoteConfig'", writeFileRule.files['.pipeline/custom_default_from_url_0.yml'])
         assertEquals('myRemoteConfig', DefaultValueCache.instance.defaultValues['custom'])
     }
+
+
+    @Test
+    void inferBuildToolMaven() {
+        helper.registerAllowedMethod('fileExists', [String.class], { s ->
+            return s == "pom.xml"
+        })
+        setupCommonPipelineEnvironment.inferBuildTool(nullScript, [inferBuildTool: true])
+        assertEquals('maven', nullScript.commonPipelineEnvironment.buildTool)
+    }
+
+    @Test
+    void inferBuildToolMTA() {
+        helper.registerAllowedMethod('fileExists', [String.class], { s ->
+            return s == "mta.yaml"
+        })
+        setupCommonPipelineEnvironment.inferBuildTool(nullScript, [inferBuildTool: true])
+        assertEquals('mta', nullScript.commonPipelineEnvironment.buildTool)
+    }
+
+    @Test
+    void inferBuildToolNpm() {
+        helper.registerAllowedMethod('fileExists', [String.class], { s ->
+            return s == "package.json"
+        })
+        setupCommonPipelineEnvironment.inferBuildTool(nullScript, [inferBuildTool: true])
+        assertEquals('npm', nullScript.commonPipelineEnvironment.buildTool)
+    }
+
+    @Test
+    void inferBuildToolNone() {
+        helper.registerAllowedMethod('fileExists', [String.class], { s ->
+            return false
+        })
+        setupCommonPipelineEnvironment.inferBuildTool(nullScript, [inferBuildTool: true])
+        assertNull(nullScript.commonPipelineEnvironment.buildTool)
+    }
+
 }
